requirements.txt
-i https://pypi.org/simple
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
flask==2.2.3
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
python-dotenv==1.0.0
werkzeug==2.2.3 ; python_version >= '3.7'
